+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## (IN PROGRESS)
 
+* Sorting of data in the "Title or package name" column is not performed in the POL list in the "Orders" app. Refs UIOR-742.
+
 ## [2.4.0](https://github.com/folio-org/ui-orders/tree/v2.4.0) (2021-06-18)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v2.3.1...v2.4.0)